Finished implementing `multiconst` macro
Bumped version to 0.2.0

Declared `multiconst` macro, which delegates all its implementation to a proc macro
(it passes `$crate` to the macro to be usable in renamed reexports)

Declared many helper items for `multiconst` macro,

Declared `FieldType` trait to query "field" types (including for tuples and arrays),
this is used when aliases are used as the type of the destructured expression.

Declared `GetFieldType` type alias for `FieldType::Type`.

Declared `Usize` type to emulate type-level integers.

Declared `__` top-level module to reexport all the items for macros.

Implemented `FieldType` for:
- all array lengths
- tuples up to 8 elements long

Note: tuples longer than 8 elements are supported syntactically
(with comma-separated types inside parentheses, not type aliases)