Added attribute support to multiconst
Now the multiconst macro allows attributes on both the passed-in const and the pattern.
The attributes above the const are copied to every generated constant,
below the attributes that were used on the identifier for the constant.

Rewrite to how patterns are parsed in multiconst's proc macro,
this will allow adding patterns that start with paths later.

Added tests for attributes.